2015 Day 22
This ended up being a bit more code and a bit more abstraction than I
wanted. However, it made the most sense in my head and it was hard to
commit to refactoring it when it basically worked first try.

Essentially, this a brute force BFS of the game state search space but
ensuring that we cut paths short if they exceed the lowest win cost.